sets are individual

+++ app/src/main/java/com/cunninghamharry/loginactivity/SetModel.kt
@@ -1,6 +1,6 @@
 package com.cunninghamharry.loginactivity
 
 data class SetModel(
-    var weight: Int,
+    var weight: Double,
     var reps: Int
 )
